ajout de visuel + voter

+++ src/Controller/PageController.php
@@ -63,6 +63,8 @@ class PageController extends AbstractController
      */
     public function edit(Request $request, Page $page): Response
     {
+        $this->denyAccessUnlessGranted('EDIT', $page);
+
         $form = $this->createForm(PageType::class, $page);
         $form->handleRequest($request);
 
@@ -85,6 +87,8 @@ class PageController extends AbstractController
      */
     public function delete(Request $request, Page $page): Response
     {
+        $this->denyAccessUnlessGranted('DELETE', $page);
+
         if ($this->isCsrfTokenValid('delete'.$page->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($page);
